de-packaged config
I didn't see the purpose to do config as a package (even
though I like the look of it when programming) it creates
a really ugly file-structure that doesn't seem correct to
me. So I'm moving config/main.go out to just config.go as
part of the main package.